devdraw: rewrite the cocoa screen
Add a new macOS cocoa screen, cocoa-screen-mojave.m.
Rewrite the macOS cocoa drawing code to use the builtin runloop,
and use the CALayer of the view with updateLayer.
Remove all of the deprecated code, and simplify some of the logic.
Modify mkwsysrules.sh such that the new code is used only when
the system version is equal or higher than 10.14.