Fix undo/redo with autoindent enabled.
Inserting a \n should commit the text, causing a following undo to backup to the \n, but not before it. However, there was a bug when autoindent was enabled that caused \n to not commit the text in most cases, allowing undo to undo beyond the \n.

Specifically in texttype, the variable r is the typed character. At the bottom of the function, if r == '\n', then it commits the text. But when autoindent was enabled, r was overwritten by the loop looking for the first non-space character of the preceeding line. Then the r == `\n` check was hit, r is the first non-space rune on the preceeding line, not the typed `\n`, and so no commit is made.